Add vim style copy & paste (+ scroll)
This removes the current scroll support (and replaces it with changing
the cursor position, meaning mouse support is still usable).
Now to enter this more we have to go into copy mode (prefix + [).
After that we can use normal hjkl movement with v for visual modes. y
yanks into the default buffer like normal, if not mose and rightclick
still is usable.