Remove tmux repeat time for keys after prefix
This means that after pressing the prefix, only one command can be
fired. This is useful, because currently when switching between two
panels that have vim or use any other vim navigation
keybindings, the tmux
"k" will interfere with the vim "k".
Example:
- In a cli panel
- Switch to a vim panel with <prefix>k
- Trying to navigate in vim with (example) j
- Instead of navigation in vim, tmux panels will be changed again back
to the previous one

So the tmux instance is "consuming" the action that was meant for vim.
This should not happen anymore, because now tmux only gets one command.