Add nerdtree for project exploring and refactoring
One of the only things I do not do in vim is firstly to look into new
projects (and might still do that in VSC). With nerdtree I have a more
feature complete filetree that i can use for that reason. The secound
and more important reason is the refactoring, renaming, deleting,
changing of files/folders. Currently i use mkdir and touch for that, but
it is cumbersome, so hopefully nerdtree will help with that.

I do not want to use nerdtree as a replacement for my fzf finding views.
This would remove too much of the vim way. And emulate some parts of
GUIs that i do not want.

If nerdtree is not providing what i want, then delete it so it does not
"poison" the current workflow with old habits from
sublime/vscode/phpstorm etc.