Reload buffer if it was changed from the outside
This is useful if I use some other tool, like vsc, in between using the
buffer in vim. Sometimes I might want to show something to others and
open vsc next to vim. When that happens i could accidentally not have
the changes I made from vsc in my already opened buffer.
This will update the buffer if something from the outside changed the
file of the buffer.