Add smooth scroll for CTRL-D and CTRL-U
This is so that CTRL-D and CTRL-U are not as visually confusing as they
are now. Currently going up or down is messing with visually
understanding it. With this plugin it will make this jump into a
scrolling animation, more like when scrolling with the mouse, meaning
that can be used instead of spamming k or j.