Disable vi mode on the terminal
This is mainly because it interacts and destroys too many other plugins
and conventions. Like reverse history search with C-R and probably alot
of others. Additionally I did not see the big positive difference
between both. I was almost 100% in insert mode, meaning that it was no
different than using the default mode.
For more complex things, I can use the shortcut to open the command line
in a vim session and edit it there with my different vim modes.