Add fzf functionality to z plugin
This will allow the z script to handle two things with fzf:
If the z command is being run without any arguments, it will open fzf
and search for all z known directories. It is like a quick search bar.
If the z command is being run with arguments, that does not return
anything normally, it will open fzf with that string in it.
A good example would be a folder name like "singularity-frontend". I
might search for it with "z sing fr" and find it. But I might also
search for it like "z sing-fr" and z does not know what I mean. This
will now open fzf and its fuzzy finding will match "sing-fr" to
"SINGularity-FRontend".