Update NERDCommenter configurations
Now it always adds the comment symbol at the start of the line. This is
really useful for auto formatting, because it will not screw up with
commented out code that is correctly indented if the lines would not
have a comment block.
Additionally now also empty lines will be commented, this is nice for
auto formatting as well but also for knowing what parts belong together.
If someone would comment out a big part, it would normally look like a
few medium sized parts. With the empty lines also commented you can see
which was commented together and what not.