Extend CR functionality in fern
When on folder, let it toggle to open and close the folder. When on file
then open it, but depending on where it is (drawer or explorer) open the
file and close the drawer (in drawer). Just open the file when in
explorer mode, because the fern "drawer" is in the same buffer.