Move plugins to the after folder
This was explained in the commit in the README, but in general this will
ensure the mappings to be mapped only after the plugins were loaded.
Otherwise it might result in a vim crash if one of them is being called
while the plugin was not loaded (eg. :Ag or fzf.vim in general).