Remove fern CR mappings
This is because the default ones are better than expected. Instead of
using ENTER to open/close folders just use h an l which is more vim like
anyway. ENTER will with that always have the same result in drawer and
explorer mode, which is probable better to keep it simple. ENTER will
always change the root path and BACKSPACE will do the same but the
opposite way.
The only thing that is missing is the automatic closing of the fern
explorer, but this wasnt handled nicely before as well. Now we have to
manually close it, but that is also more like the way how a tree
explorer would work. Instead of having muddled philosophies every
different way of opening a file uses their own way of thinking:
- buffer shows only current layer and changes root with ENTER BACKSPACE
- tree explorer is "always" visible and shows whole project (h,j,k,l,)
- searching with fzf.vim etc will handle everything with regex