Extend Ag command to also show hidden files
This is mainly useful for env files but also for this current project as
a whole. Most of the files are in a hidden folder so without this it is
not easily possible to search through them for something like mappings
or configurations without knowing where to look or randomly searching.