Ignore spellings that were added with zg or zw
At least for now, maybe in the feature it is a good idea to have a
central list of words that I still want to use, but currently it seems
to be just annoying to always keep it up to date and commit it. This
repo does not need more unnecessary commits.