Update plugin/mapping/config structure
Now the after folder is only used for things that really only need to be
loaded after everything like mappings (because they could depend on
plugins to run).
Normal configurations are in the normal plugin folder that is being
loaded with the normal init.vim instead of afterwards like the after
folder.