Set indent foldmethod as default but do not show
For vim always use the marker method and always fold on open buffer.
With this I should have the best of both worlds, possibility to fold on
any project I want without it being annoying (I hope and will see) and
my vim files always be folded with the marker in mind (hope that does
not have bad consequences).