Update window mappings to be similar to others
Tmux and other general settings have some standard, and this should
mimic it a bit or otherwise do not use other things that might conflict
with each other.
As an example, to get a full screen (zoom in) of a window fern, tmux and
others use the "z" key. My settings used the "f" key what makes sense
for "fullscreen" but was confusing to have two words that mean the same
on two different keys.
Another exmaple would be "c" for tmux create and here for window close. So
the opposite meaning but the same key, because of that and because there
is no real good mappings, I just removed it (mostly use :q either way).

Others are to equalize the windows, "e" could be referenced for
something else like editing, so remapped it to "=" and only with "o"
could mean something like only but also open, so "1" is now being used
for it.