Remove undotree plugin
I almost never used it. Almost every time I opened it, I did it by
accident. In the whole time I used it, I only used it once or twice in
practice. So because of it being used rarely I want to remove it and see
how it works without.