Update call of compinit
Before I just had it pasted in it, because some plugins (git, docker)
did not work. As far as I can remember it was the completions.
After reading into it a bit more, and finding out that this is the main
reason why my shell loaded that slow update the call to it, to just call
it with the autoload builtin.
This will generate a .zcompdump that can be seen as a cache of
completions. And result in a faster bootup.

All in all I got to the current state:
- ~1s all plugins and buggy compinit
- 0.1s minimal plugins and buggy compinit
- >0.1s all plugins and fixed compinit

As visible, the compinit was the main problem, so now I can add plugins
as I want again without needing to worry about startup performance.