Show git hash if HEAD is detached
Ignore if the HEAD is empty, this will now just show an empty value on
the statusline, but I think that should be fine.
Firstly I rarely am not in a git repo with vim and secondly it does not
look too bad without needing to invest too much time into fixing it.