Add alignment to mode statusline
This means that now the part of the statusline that shows the mode and
detailed mode will not jump around as much as before. Some modes still
have that effect but by far the normal workflow does not jump around
anymore (insert, insert completion, normal).
Mode is aligned to the right and detailed mode is aligned to the left so
that it looks like both have paddings on the outside.