Fix true color mode for tmux
The problem was, that the terminal emulator needs to be setup correctly
as well. Meaning that we need the TERM env var to be also a true color
one. Before it was set to "alacritty" by default, which worked without
tmux but as soon as tmux was used, it caused problems.
Now we set the TERM env to xterm-256color.

This could also be set inside of alacritty, but I think it makes more
sense in the zshrc instead.