Remove clever-f plugin
Instead of using f and a key for more complex searches, just use /
instead. So if I have something like:
"Today is a really great day"

And I want to get to the "a" in day. Instead on relying on f and using
clever-f highlighting to easily navigate to it, just use "/ay" and n.