Let the terminal decide on the TERM env
This is mainly because kitty should use its own ENV for correctly
representing itself.
This fixes stuff like undercurl etc, but it will mean that others might
also now know that kitty is being used and might not allow it.
For example tmux on the server does not know tmux and will not allow it
because of that.