Remove dynamic status line color change
The execute of the highlights caused vim to slow down by quite a lot. To
prevent that I tried doing it in another way, but wasnt able to find a
nice and easy to understand solution.
In general the different colors are just something that is nice to have,
the name of the mode is still being displayed which is more than enough
for me personally.

If I ever find a good way of handling the change of the colors, I might
revisit this issue again (maybe with events or with lua).