Set file dependent vim options for kitty/vimrc
This is once for kitty for readability, meaning putting it on the first
line, that people can see if and not overread it (being confused why it
works).
For init.vim it is because it is really only needed for this file in
general, before it was added for all .vim files what is not really
needed.
Like this it also works with sourcing, meaning now when changing
something and sourcing it, it will also apply the correct foldmethod.
Before it didnt, and was overwritten by the default in the init.vim.