Remove specific man pager
This is because it looks nice with colors, but sadly it will result in
other problems.
One of them is that the usage of the mouse wheel is not working properly
anymore and there might be other problems as well. Because of that ill
keep it at the basic usage.