Remove not needed gitignore
Now only the needed files are being linked, meaning that this repo does
not have all of the dotfiles in it, but only the referenced ones.
Because of that we can remove most ignore rules again. This makes other
things like "ag" in nvim also better (search will find dotfiles).