Replace fern with nvim-tree
This alternative is faster and more "vim-like" meaning that other
commands like "/" or "n" still work in that buffer.
Additionally it does not need other plugins to work independently with
git.
It also has some drawbacks. Gitignore files are not correctly shown,
this might be a problem in the future but for now it should be fine.
Another one is just an inconvenience. Now I need to cut and paste to
move files instead of doing it by the file, its a pro and a con, mainly
new muscle memory