Outsource bighost command to init.vim
This is mainly because before I did not want it in the main config file
of fear that I might accidentally press the mapping and upload random
data to the server.
Now it is just a command that I need to actively run (similar to a
plugin) so this problem should not be there anymore.
If I need a map for quick upload because I have to run it over and over,
I can just create one on the fly as I think I should in that instance.