Remove rofi
I rarely used rofi as a window switcher because firstly I do not really
need a specific switcher. Either I use the os built in ones to swtich
between virtual desktops or I just click on the needed application if
its a special one.

Because of that Ill delete rofi completely, but might install it again
in the future if this feature is not present in other distros etc.