Remove fugitive
This is mainly to try out lazygit bit more without falling back to
fugitive. Fugitive is really good but I still see myself changing back
to the basic git commands and to vscode for branches and merge
conflicts. I also want to have less overhead in my init.vim, the git
workflow can be extracted out of vim.

Still it might that Ill return to fugitive because it is a really nice
plugin.