Try out telescope as an fzf alternative
This is still in testing period, but it already has some nice features
that I would miss when looking at fzf.
I can configure and add almost anything I want as well.
Only visible problem so far is the speed. Because of that Ill leave fzf
in my config until I see that I dont need it anymore.