Comment out fzf config
This is only a temporary thing. It should not stay in my config for too
long. It is just needed, because my telescope mappings would conflict
with fzf. I also want to use the same, so that I can get used to the new
plugin without mentally using it.
As the previous commit said, this is just until im sure if telescope is
good enough or if there are any dealbreakers.