Remove fzf from vim
This will be replaced by telescope.nvim. It has many other things I
quite like about and would miss in fzf.vim.
Searching gigantic repos is pretty bad with telescope, but might be
fixed in the future and with fzf it was never the best possible. I
should try to use manual ag/rg/fd instead of integrated in vim. That
might be better and by far faster.