Remove v and vv alias
This is because I mainly used "vv" meaning "nvim .", but this will
always add the repo into the argslist. So "nvim" is a better command to
run in the end. Because of that there is no real reason to add an alias
that I use often, but not often enough to go from "nvim" to "v" mainly
because "vim" is mapped to "nvim".
So in the end I rather use "vim" instead of "v" and remove these maps
entirely.