Add plugin "which-key"
This will show the mappings while typing them. Normally that would not
be needed, because I should know my mappings and others that I do not
use I should remove.
But with the lsp, we have some mappings that I need but rarely use and
now I tried unifying the mappings (also easier to showcase with plugin)
so for some time I might need the hint, so that I do not need to look
into the config file.

All in all it might be that this plugin is just temporary and will be
removed later on when I am more happy with my overall setup.