Remove "non essential" plugins and mappings
TLDR: Bring my config to a minimal point, which only has vim specific
things, and does not try to emulate other things
that I dont really need in it.

These include:
- git-messenger (git-blame)
- vim-floaterm (for git commit)
- nvim-tree (file explorer, replaced by netrw)
- nvim-web-devicons (icons for telescope/nvim-tree)
- vim-which-key (show mappings)
- nvim-colorizer (show colors in buffer)
- mappings for resize and pasting

These are all things that are nice to have, but rarely used.
Now that I am trying to get more vim things out of it and mainly use
vim as a terminal text editor instead of an IDE,
I want to get back to a more minimalistic config.

This is considered for the future, if I want to go to another tool,
which has a basic vim-emulation or neovim embedded functionality
then I want to get only things from vim, which
are from vim specifically.

So want to remove things, that I try to emulate into vim
which are not really vim, or are just something that takes it too far.

So:
- dont need git-blame in vim, can just use vscode or the browser or git
- do not need a terminal in vim, can just use a terminal
- do not need a fancy file tree, have telescope/fzf and netrw
- do not need fancy icons which do not work correctly most of the time
- do not need showcase of mappings, should have some and know them
- do not need colors in vim, rarely use them either way