Update information about vscode pluins
Remove script for automatic installing, because only one of these
plugins will most likely be used (vim). For that one add a description
why to use it and the repo for later reference.

The others should not be included everytime, if I find good plugins I
want to keep, I can add them in the readme but not in a script that
installs them without me really noticing.