Remove vscode if block from statusline plugin
This was used to prevent the neovim vscode plugin to call it.
It should have been removed with the commit that removed that plugin.
This should be found a few commits ahead if needed.

VsCodeVim seems to not call the standard neovim directories
so I dont need to have any edgecases for it.