Handle special packages differently in ansible
Instead of making it more general, I hardcoded the fnm path and
take as a given that fdfind is installed, so that ansible knows
if these calls are the same or updated.

Before every call to the "which" part was consideret as "updated"
meaning ansible thought this process was always new.