Split all sub packages into their own import
Using the loop does not allow to add a tag to the included tasks.
I want to have an easy way of only calling specific parts of packages,
later on this might be fixed, by putting everything into their own
roles but for now with this everything can be used independently.

Was trying to find a way to abstract it, with templates, modules,
tasks with only variables.
But everything did not work or was way too unreadably for the benefits.