Install vscode plugins not as root
For that I could also set it as true per default and overwrite it to
false for the plugins. I like it better to do it explicitly, meaning set
it to true to every needed place (only one more item to add compared to
the other way).

Also update the correct path to the cute binary.