Remove tags from ansible
When executing the playbook with tags, this will run it without
gathering facts. These facts are really useful to get properties like
the current system etc.
Because I want to use them and do not want to be confused why it
sometimes does not work, ill remove the tags, and optimize the scripts
more for speed, so that it should not be slow when stuff is already
installed (mainly basic packages).