Improve config for quickfix usage (eslint/tsc/ag)
Use :Ag so that we can use its power for quickfix:
- :Ag [pattern]
- all results are in fzf buffer
- filter in fzf buffer
- <tab> or <ALT-A> to select multiple entries
- <enter> to put selection into quickfix

Additionally add testing commands for eslint and tsc.
Both just run a command in the shell, and pipe them through sed/awk to
get a format back which can be understand by the quickfix format.

These are mainly added to git, so that I have a reference of them in the
future (unless I keep using them then they are here to stay).