Hide mode and disable pair highlight
Showing the mode was nice, when I was trying to use different completion
modes. Now I am rarely using them and if I am, then I know them and will
not look at the screen.

The other option will highlight the open bracket. This does not really
do anything, because my lsp already does that. And it is a bit
  distracting (it moves the cursor to highlight it).