Change vim hop mappings to leader s from f
Main reason is, that hop does not really allow reproducible actions.
Meaning if I write a small macro and I want to find ( then this will not
really work with Hop and I jump around everywhere.

Because of that I am binding it to the leader s key, which is a bit
annoying to handle, but should remove breaking vim like bindings/ways.