Install most treesitter languages and disable some
Before I installed some of the languages that I normally used and
ignored others where the highlighting is still lacking (compared to
regex).
But also other features are useable without the colors. Context,
Objects, etc. are being used by plugins making the editing experience
better.

Example is php, the color is not that great but showcasing in what
context I am currently (what function, etc) works really well in most
php projects (that have gigantic functions etc).

So now instead of having to note down what languages not to use, I just
install all of them (that are maintained) and just disable the highlight
of languages, that I do not like (mainly php and vim).