Remove automatic installation of VSCode
Ansible was used to install vscode and all its plugins beforehand. This
worked mostly but was almost everytime the painpoint of the whole
config.
Either the installation was slow, because vscode is quite big or it even
broke my apt command because of some weird keys that semi-work.

I am not planning on using vscode as my text editor of choice, so I do
not really need to automatically install it. Installing it via a gui
when I want to should be better, also the only real benefit was the
installation of my plugins, which is not really difficult, meaning just
writing down my favourites and manually installing them when needed
should be more than enough.