Add quickfix command for jest tests
This is not really independent, because I need to prepend "src" to the
endresult sadly.
Jest does not give the full path, and there is no easy or short way of
doing it. I could use the --json flag, but then I need to parse the
whole json part with jq, remove parts a few times with sed and cut other
parts with cut. Resulting in a gigantic oneliner that noone will get.

Because of that I use the not optimal but shorter oneliner.