Hide AWS in command line
The information will not change for me and having it all the time on the
command line is not really helping.
I am never really logged out as well, and if I am then I will notice
(hopefully), so having just a symbol also does not really do anything
for me.