Remove lspsaga
Lspsaga does not allow me to jump into the floating window (at least I
cant find how). Additionally the features I get from it are either buggy
or not as useful.
Two main things I use them for:
1. rename
2. code action

For renaming, when moving in floating window, the buffer bugs and I can
not type anymore (mostly when moving in the buffer).
for code action, it only works for actions under the cursor (not buffer
wide) and that is almost the same as native.

So the only real benefits it gives are styling (looks way better than
native) and fancy floating windows. I think I should just remove it and
search for another alternative in the future or make something myself.