Add null-ls as a formatting tool
The only thing that I am missing is an easy way of formatting a buffer.
It is possible to do it externally or call cli programs, but the LSP has
a formatting endpoint which I would like to use for that.
null-ls allows to hook into this formatting part and uses different
formatters for that. Now I can use prettier for quick formatting which
looks to me like the LSP does all the work (which it kinda does, because
now null-ls is a lsp that uses prettier etc.).