Remove tagfunc/formatexp/omnifunc for lsp in vim
I never really use it. Maybe the formatting, but normally I just format
the whole document anyway (maybe it differs in the future).

It is too much of an ugly setting to keep for no benefit. I want to keep
my config somewhat reasonable and this just seems to be too much for
nothing.