Hide swap-file warnings in vim
I think I still want to keep the swapfile, because I had it once that my
vim crashed and lost all the changes. Normally these are not many, but
if I can get a "backup" for free why not.
Sadly swapfiles have a lot of false positives, meaning I get the warning
often, if I open the file multiple times.
To have the best of two worlds (open in multiple sessions without hassle
and recovering in case something happened), I just ignore the error and
manually run ":recover" in case I notice that data is missing.

Only thing that might need to be seen is, if this makes problems with
larger files, if this is the case, then I should deactivate swapfiles
(noswapfile) because the benefit is not really big and kinda not needed
because of git and my general workflow of manual saving.