Update nvim-dap config/mappings
Put the debuggers into the xdg data folder (mostly .local/share) instead
of the home directory, for better usability.
Also add a mapping to break on exception as this might be a good thing
to have and remove the repl mapping, because the UI is already giving me
that in full, and the mapping is just toggling the visibility, resulting
in me accidentally closing the repl.