Add buffer mappings/configs for help/lspinfo
The new mapping in lspinfo will allow to use 'q' to quit the window
(similar to most other floating windows).
The config in help will allow for easier writing of help files, because
the special symbols will be visible and the cursor gets highlighted
better for easier formatting and alignment.