Deactivate matchit plugin
It sadly does not really work well with visual mode. Meaning to jump
from one part with % to another to update a visual selection.

The main problem is in the global command, v% works in a normal scenario
and in a macro, but as soon as we call the macro or the normal commands
inside a global command, unexpected things happen.
It seems like the start position does not get it from global but from
where my cursor was when running the command

e.g.:
  :g/bar/norm nf(v%d

for:
  foo.bar('example')

It should get all references to bar, jump to bar then find open "("
start visual selection to jump to the closing ")" and delete it.
This works when matchit is not active, if it is active then everything
from the cursor until the last bar + norm will be deleted (or similar,
in general it does not what it should)