Remove rarely/never used lsp maps
As far as I remember I only use jump to definition and sometimes jump to
type definition. The other two are not really used, and mainly do the
same as the two that I mentioned.
gi/gI are replacing useful vim mappings, that I might use (not often but
more often than the other two lsp mappings). So I updated to only use
gd/gD for my goto definition (even makes more sense for the names) and
get back the gi/gI.