Use folding with treesitter
Instead of deactivating the folding per default and using indent, I now
always have it active, but with a high fold start level.
This is mainly done, so that the first closing of the folds will not
start at level 0, which would result in everything getting closed.

Additionally, using the foldlevels is really nice, I can use `zR` to set
it to the highest in the current window, and wit `zm` I can decrease the
level, so all scopes that are in a level are closed

Example
```
function foo() {
  if (condition) {
    example
  }

  if (condition) {
    example
  }

  example
}
```

will show only the things directly in the function, if the foldlevel is
low enough
```
function foo() {
  if (condition) { hide }

  if (condition) { hide }

  example
}
```