Remove gpg signing per default
The reason behind it is, that all my future commits will be signed by
multiple gpg keys on multiple computers. These computers or
installations might change rapidly, resulting in short lived keys.

To keep all commits verified I would need to add every key and every
passphrase, even if the private key does not exist anymore, just so that
my verified status of the commit can stay.

Because the main part for trying out gpg was this dotfiles repo and I
commit to it with multiple computers, it is just not worth it, to
"trash" my github keys full of past keys.

In the future and on specific more open and public repos, I might will
add gpg signing again, and even more if I know that the computer stays
or I find a good way of storing the keys.