Decrease tmux history limit
Higher limits are nice for long backtrack, but if the whole buffer is
filled with information, too many lines will result in the buffer to
lag, resulting in resizing and scrolling to be unresponsive and even
freeze the application.

Decreasing it a bit, results in less lag and most of the time I do not
need to go that far into the history either way.