Remove debugger (vimspector) in vim
I am not really happy with the debugging experience in vim with
vimspector (and also with the native one of nvim with dap). So instead
of having a big python dependency which I never really use, I just
remove it and wait until there is a better solution that fits my needs.