Remove ansible
Ansible works fine, but I have some problems with it:
- it is gigantic and i dont use any of it
- big dependency
- a lot of boilerplate
- need to learn ansible
  - checks, env vars, etc. need to be translated into ansible
  - this is way easier in basic bash scripts
- never really sure if it works, as it is a lot of magic

Because of these and probably other reasons, I tried writing some small
scripts which do 98% of what my current ansible setup does.
The scripts are not short, but generally the boilerplate of 30 lines can
be shortened into 5 bash script lines.

So I decided to switch to a more bare bones solution.