Rebind tmux pane editing shortcut to `e`
The reason is, that `S` is firstly overlapping with `s`, it needs
`shift` and that other environments have already somewhat of a standard.

In fish `alt-e` will open command line in vim and in other shells, some
combination of `ctrl` and `e` do something similar.

So to keep with the tradition and muscle memory, I rebound it from `S`
to `e`.