Show wget output in install script
The main reason is, if we do not show the output, then it could happen
that the server is slow and we will not see any output in the install at
all.
This results in confusion, and it looks like the script is broken (just
because the server is not available or we need to wait a bit until the
next retry).