Update tmux scratch session to a note session
scratch was meant for running throwaway commands (like aws login or
tunnels), as these happen rarely, the scratch buffer was not really used.

Now with my try of using vim also for notetaking, this session makes
more sense for the notes and can still be used for something of a
scratch session (in a new tab or pane).