Use vscodes 3-way merge for complex conflicts
I can not find a good cli tool that helps with complex merge conflicts.
Most vim tools are either without any syntax highlight or are really
basic and show all changes in one buffer (so not much more then just
opening the conflict file by itself).

VsCode just got an update which adds an improved 3-way merge option. I
am not really using vsc, but want to try using this mergetool until I
find a better option.

A drawback from this solution is, that git, mergetool or
vsc for some reason creates the different variants of files of the merge
conflict. Meaning it creates files for BASE, LOCAL, REMOTE which spam
the folder full of not needed files, which need to be deleted
afterwards.
I was not able to find a way to prevent them, as git and mergetool do
not say anything about not generating them. As far as I can see these
files are for compare buffers (which is not needed for vscode).