Replace fzf-lua with telescope
The main reason, I did not want to use telescope was, that it is slow in
really big projects, and that it does not have the fzf syntax to filter
stuff out (!ignore 'exact-match etc).

Firstly I do not have that gigantic projects currently, so that is not a
big concern anymore. Secondly I use a fzf clone for telescope, which
should also be able to handle big projects and supports fzf syntax.

Another big reason for telescope is, that it integrates nicely with a
lot of other neovim things. Mainly lsp, treesitter, etc.