Replace coc with native lsp
Coc has way more out of the box, but it also does not have the nicest
integration with newer neovim features or extendability, unless things
are written for the coc plugin system.

Both coc and native lsp have pros and cons, so there is no reason to
switch between each other from time to time.

Some of the reasons I didnt stick with native lsp last time, was that
the main completion plugin (cmp) screw up macros, that the lsp did not
attach correctly to the buffer, that some processes took way too much
memory while running and generally a way too cumbersome setup.

Some parts are resolved some are improved. Cmp doesnt destroy macros
anymore (works similar - not as good as coc - to coc). Lsp seems to be
auto attaching way better (did not see any problems so far). Processes
take good memory, but not sure if that differs to coc, as both use the
same lsp in the end. Setup is easier thanks to mason and general
improvements.

The main reason to switch is to try it out (as last time was about
half, maybe more, year ago). And use the better integration to neovim,
treesitter and the plugin system (telescope etc).