Add dap for debbuging
As I still want to try out a debugger and I now use mason, which also
handles debuggers (which was annoying to install previously - with
external scripts etc), I want to give it another try.

Additionally it got improvements, and the community is starting to
create better and better plugins for the dap. So in the future it might
be improved even more after lsp is implemented well enough and focus is
set to other places.

For installing, mason needs to be run manually and for continuing the
debugger, I want to try out the repl, which has commands to continue
etc.