Replace gruvbox with onedark colorscheme
I really like gruvbox, but sadly it does not include correct colors for
diffs. This results in all git diff plugins and functions to show
added/removed text just as a big green or red blob with no syntax
highlight.

I want to try to use vim more with git, as some parts of lazygit do not
correctly work or are not as integrated into vim as I want them to be
(and some parts I do on the commandline either way already).

Additionally I really like the onedark warmer colorscheme, so I want to
try out something new.