Switch from init.vim to init.lua
There is no real reason for this switch.

As my plugins are more and more lua based, I wanted to see how it feels
to use lua instead of vim as a configuration language.

It might be more complex, as I have multiple configs at different
places, but having a linter, formatter, lsp is really nice to keep the
configs etc consistent.

Also it makes integration of more complex functions and commands (like
my indent text object) way easier to implement, test and use.

It had some small boost in performance, but nothing impactful.

One negative point is, that if one part breaks, possibly my whole setup
breaks (as it might throw instead of ignoring the errors) which is
possibly not a bad thing, as this will show that something is broken and
not just let me continue until i stumble upon that problem myself.