Remove lsp_lines
Mainly because it is more bothersome than helpful. In rust it spams the
whole buffer full of messages, and I mostly need to hide it with my map
to toggle the visual effect.

Similar to ts, as I dont need a big benefit, ill remove it again and use
diagnostic lines (native from neovim).