Replace vim cursor restore audocmd with lua
It seems like the filetype detection does not work currently anymore
(not sure if it is because of WSL or treesitter or something else).
This means that I can not check if the filetype is a gitcommit or
something else, where I dont want to jump to a different location.

But this is just a small thing (and seems to not be working before
either way, so will fix it in the future.

Just as a note:
use `vim.o.filetype` and `string.find(filetype, 'commit')` to quickly
write lua equivalent to the vim part which I remove with this commit.