Remove delta from lazygit because of bug
For some reason my lazygit has problems with delta output.
The first problem is, that it will just not show the how diff inside of
lazygit and even worse, the staged changes most of the time are
completly missing.
When refreshing the pane, it shows it sometimes again but it vanishes
afterwards again. Resulting in it to be almost unusable.

Sadly delta never worked in manual hunk staging.

Lastly there are some cases where lazygit just crashes because of the
diff.

So I went back to the basic git diff instead, to see how bad it is and
if it makes any problems. Could try other diff views, but didnt really
see others that I like as much (or more than just the basic).