Remove diffview for vim
I tried integrating more git plugins to enable easier merge resolve in
vim itself.
But it did not really work as well (mainly because I rarely need to use
it and when I need it, I dont know about the keymaps etc.).

So in the end I did not use it and went back to lazygit for 95% of the
stuff, changing into vim for easy conflicts and use `git mergetool`
which is linked to vscode for more difficult conflicts (as a visual gui
still is easier for me personally).