Remove dap/debugger from vim
Sadly (similar to git integration) I did not really use dap at all. When
I used it, I did not know about the keybinds as I dont use it often
enough.
Also it is still quite buggy (at least for js), it jumps to compiled
places for promises, which pollute my jumplist and are just annoying to
navigate through.
Finding values does not work as easily as it should and
starting/attaching/stopping the dap does not work as seamlessly as I
hoped if would.

As an alternative I want to use an external tool for this, probably
vscode as a test and see later on.

I might come back to this when it is more mature as this would be quite
nice inside of my main editor.