Use delta for lazygit
I removed it in the past as I had displaying problems. This seems to be
still the case (but rarely). As I use lazygit mainly to view my changes,
having it not working sometimes might be better than having worse
highlights.

So this is mainly as a tryout between basic git for extreme things,
fugitive for some smaller things and lazygit for visual things.