Replace packer with new lazy package manager
One reason to do this, is that I always wanted to use more lazy loading
and this package manager does it in a nice way that I understood
(compare to packer where I had no idea what happened).

Another reason is, that this also allows for easy split of the configs
in different files, as it will merge them together in the end, and was
build with hindsight of using it split into multiple files (thanks folke).

The UI is way better and allows for easy profiling, testing.
It is similar to mason so a general UI with different tabs.

It uses a lock file instead of needing to manually locking the versions,
so I can go back to my previous working version without any need to do
anything special before.
No need to also manually compile the plugins to get the performance
benefit (packer) or forgetting it and having outdated plugins without
knowing it.

Lazy installs on open, meaning I do not need to remember this on new
install.
And it checks for updates (not sure if thats good or bad, will see).

Performance is already included, meaning I dont need the impatient
plugin (for packer).